Finna API support for Ajapaik Android

+++ project/ajapaik/api.py
@@ -19,7 +19,7 @@ from django.contrib.gis.measure import D
 from django.contrib.sessions.models import Session
 from django.core.exceptions import ObjectDoesNotExist
 from django.core.files.uploadedfile import InMemoryUploadedFile
-from django.db.models import Q
+from django.db.models import Q, Count, Case, When, Value
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
 from django.utils.translation import activate
@@ -37,10 +37,20 @@ from rest_framework.views import APIView, exception_handler
 from sorl.thumbnail import get_thumbnail
 
 from project.ajapaik import forms, serializers
-from project.ajapaik.models import Album, Licence, Photo, PhotoLike, Profile, GeoTag
+from project.ajapaik.models import Album, Licence, Photo, PhotoLike, Profile, GeoTag, Source
 from project.ajapaik.settings import (API_DEFAULT_NEARBY_MAX_PHOTOS,
                                       API_DEFAULT_NEARBY_PHOTOS_RANGE,
                                       GOOGLE_CLIENT_ID)
+
+from project.ajapaik.curator_drivers.finna import FinnaDriver
+from requests import get
+import re
+import sys
+from django.core.files.base import ContentFile
+
+
+from project.ajapaik.curator_drivers.finna import find_finna_photo_by_url, import_finna_photo
+
 
 log = logging.getLogger(__name__)
 
@@ -358,6 +368,100 @@ class AlbumList(CustomAuthenticationMixin, CustomParsersMixin, APIView):
             ).data
         })
 
+class FinnaNearestPhotos(CustomAuthenticationMixin, CustomParsersMixin, APIView):
+    '''
+    API endpoint to retrieve album photos(if album is specified else just
+    photos) in specified radius.
+    '''
+
+    permission_classes = (AllowAny,)
+    search_url = 'https://api.finna.fi/api/v1/search'
+    page_size = 30
+
+    def post(self, request, format=None):
+        form = forms.ApiFinnaNearestPhotosForm(request.data)
+
+
+        if form.is_valid():
+            lon=round(form.cleaned_data["longitude"], 4)
+            lat=round(form.cleaned_data["latitude"], 4)
+
+
+            finna_result=get(self.search_url, {
+#               'lookfor': cleaned_data['fullSearch'],
+#               'type': 'AllFields',
+                'limit': self.page_size,
+                'lng': 'en-gb',
+                'streetsearch' : 1,
+                'field[]': ['id', 'title', 'images', 'imageRights', 'authors', 'source', 'geoLocations', 'recordPage',
+                        'year', 'summary', 'rawData'],
+                'filter[]': [
+                    'free_online_boolean:"1"',
+                    '~format:"0/Place/"',
+                    '~format:"0/Image/"',
+                    '~usage_rights_str_mv:"usage_B"',
+                    '{!geofilt sfield=location_geo pt=%f,%f d=0.1}' % (lat, lon)
+                ],
+                })
+
+            photos=[]
+            results=finna_result.json();
+            for p in results['records']:
+                comma=', '
+                authors=[]
+                if 'authors' in p:
+                    if p['authors']['primary']:
+                        for k,each in p['authors']['primary'].items():
+                            authors.append(k)
+
+                if 'geoLocations' in p:
+                    for each in p['geoLocations']:
+                        if 'POINT' in each:
+                            point_parts = each.split(' ')
+                            lon = point_parts[0][6:]
+                            lat = point_parts[1][:-1]
+                            p['longitude'] = lon
+                            p['latitude'] = lat
+                            break
+
+                ir_description=''
+                image_rights=p.get('imageRights', None)
+                if image_rights :
+                    ir_description=image_rights.get('description')
+
+                photo= {
+                    'id':'https://www.finna.fi%s' % p.get('recordPage', None),
+                    'image':'https://www.finna.fi/Cover/Show?id=%s' % p.get('id', None) ,
+                    'height':768,
+                    'width':583,
+                    'title':p.get('title', None),
+                    'date':p.get('year', None),
+                    'author':comma.join(authors),
+                    'source': {
+                                  'url': 'https://www.finna.fi%s' % p.get('recordPage', None),
+                                  'name': ir_description
+                              },
+                    'azimuth':None,
+                    'rephotos':[],
+                    'favorited':False
+                }
+                photos.append(photo)
+
+
+            return Response({
+                'error': RESPONSE_STATUSES['OK'],
+#               'photos': finna_result.json()
+                'photos': photos
+            })
+        else:
+            return Response({
+                'error': RESPONSE_STATUSES['INVALID_PARAMETERS'],
+                'photos': []
+            })
+
+
+        return HttpResponse(response, content_type="application/json")
+
 
 class AlbumNearestPhotos(CustomAuthenticationMixin, CustomParsersMixin, APIView):
     '''
@@ -368,6 +472,10 @@ class AlbumNearestPhotos(CustomAuthenticationMixin, CustomParsersMixin, APIView)
     permission_classes = (AllowAny,)
 
     def post(self, request, format=None):
+        if request.user:
+            user_profile = request.user.profile or None
+        else:
+            user_profile =  User.objects.get(pk=17345994).profile
         form = forms.ApiAlbumNearestPhotosForm(request.data)
         if form.is_valid():
             album = form.cleaned_data["id"]
@@ -393,6 +501,10 @@ class AlbumNearestPhotos(CustomAuthenticationMixin, CustomParsersMixin, APIView)
                     .distance(ref_location) \
                     .order_by('distance')[start:end]
 
+#                    .filter(
+#                        Q(likes__profile=user_profile) | Q(likes__isnull=True)
+#                    ) \
+
                 return Response({
                     'error': RESPONSE_STATUSES['OK'],
                     'photos': serializers.AlbumSerializer(
@@ -412,7 +524,7 @@ class AlbumNearestPhotos(CustomAuthenticationMixin, CustomParsersMixin, APIView)
                         geography__distance_lte=(ref_location, D(m=nearby_range))
                     ) \
                     .distance(ref_location) \
-                    .order_by('distance')[start:end]
+                    .order_by('distance')[start:end] \
 
                 photos = serializers.PhotoSerializer.annotate_photos(
                     photos,
@@ -498,10 +610,29 @@ class RephotoUpload(CustomAuthenticationMixin, CustomParsersMixin, APIView):
     '''
 
     def post(self, request, format=None):
+        print >>sys.stderr, ('rephotoupload')
         form = forms.ApiPhotoUploadForm(request.data, request.FILES)
         if form.is_valid():
             user_profile = request.user.profile
-            original_photo = form.cleaned_data['id']
+
+            print >>sys.stderr, ('form.isvalid()')
+            id = form.cleaned_data['id']
+            if id.isdigit():
+                id = int(id)
+                photo = Photo.objects.filter(
+                    pk=id
+                ).first()
+            else:
+                photo = find_finna_photo_by_url(id, user_profile)
+
+            if not photo:
+                print >>sys.stderr, ('rephotoupload failed')
+                return Response({
+                    'error': RESPONSE_STATUSES['INVALID_PARAMETERS'],
+                })
+            
+ 
+            original_photo = photo
             latitude = form.cleaned_data['latitude']
             longitude = form.cleaned_data['longitude']
             rephoto = form.cleaned_data['original']
@@ -560,6 +691,7 @@ class RephotoUpload(CustomAuthenticationMixin, CustomParsersMixin, APIView):
             image.save(image_stream, 'JPEG', quality=95)
             image_unscaled.save(image_unscaled_stream, 'JPEG', quality=95)
 
+            print >>sys.stderr, ('new_rephoto')
             new_rephoto = Photo(
                 image_unscaled=InMemoryUploadedFile(
                     image_unscaled_stream, None, rephoto.name, 'image/jpeg',
@@ -585,6 +717,7 @@ class RephotoUpload(CustomAuthenticationMixin, CustomParsersMixin, APIView):
                 user=user_profile,
             )
             new_rephoto.save()
+            print >>sys.stderr, ('original_photo')
 
             original_photo.latest_rephoto = new_rephoto.created
             if not original_photo.first_rephoto:
@@ -693,35 +826,51 @@ class ToggleUserFavoritePhoto(CustomAuthenticationMixin, CustomParsersMixin, API
 
     def post(self, request, format=None):
         form = forms.ApiToggleFavoritePhotoForm(request.data)
-        if form.is_valid():
-            photo = form.cleaned_data['id']
-            is_favorited = form.cleaned_data['favorited']
-            user_profile = request.user.profile
+        
+        photo = None
 
-            if is_favorited:
-                try:
-                    # User already have this photo in favorites.
-                    # Nothing to do here.
-                    PhotoLike.objects.get(photo=photo, profile=user_profile)
-                except PhotoLike.DoesNotExist:
-                    # User haven't this photo in favorites.
-                    # Creating like.
-                    photo_like = PhotoLike.objects.create(
-                        photo=photo, profile=user_profile
-                    )
-                    photo_like.save()
+        if form.is_valid():
+            user_profile = request.user.profile
+            id = form.cleaned_data['id']
+
+            if id.isdigit():
+                id = int(id)
+                photo = Photo.objects.filter(
+                    pk=id
+                ).first()
             else:
-                try:
-                    # User already have this photo in favorites.
-                    # Removing it.
-                    photo_like = PhotoLike.objects.get(
-                        photo=photo, profile=user_profile
-                    )
-                    photo_like.delete()
-                except PhotoLike.DoesNotExist:
-                    # User haven't this photo in favorites.
-                    # Nothing to do here.
-                    pass
+                photo = find_finna_photo_by_url(id, user_profile)
+
+            if photo :
+                is_favorited = form.cleaned_data['favorited']
+
+                if is_favorited:
+                    try:
+                        # User already have this photo in favorites.
+                        # Nothing to do here.
+                        PhotoLike.objects.get(photo=photo, profile=user_profile)
+
+                    except PhotoLike.DoesNotExist:
+                        # User haven't this photo in favorites.
+                        # Creating like.
+                        photo_like = PhotoLike.objects.create(
+                            photo=photo, profile=user_profile
+                        )
+                        photo_like.save()
+                else:
+                    try:
+                        # User already have this photo in favorites.
+                        # Removing it.
+                        photo_like = PhotoLike.objects.get(
+                            photo=photo, profile=user_profile
+                        )
+                        photo_like.delete()
+                    except PhotoLike.DoesNotExist:
+                        # User haven't this photo in favorites.
+                        # Nothing to do here.
+                        pass
+
+
             return Response({'error': RESPONSE_STATUSES['OK']})
         else:
             return Response({'error': RESPONSE_STATUSES['INVALID_PARAMETERS']})
@@ -770,6 +919,8 @@ class PhotosSearch(CustomAuthenticationMixin, CustomParsersMixin, APIView):
     '''
     API endpoint to search for photos by given phrase.
     '''
+
+    permission_classes = (AllowAny,)
 
     def post(self, request, format=None):
         form = forms.ApiPhotoSearchForm(request.data)
